requirements.txt
# Inventory Management System Requirements
# Install with: pip install -r requirements.txt

mysql-connector-python>=8.0.33
